Add basic FFI go grammar.
Prefixing a function or variable definition with the keyword
"foreign" declares a function or variable without defining it.
The right-hand side of "=" should be a name rather than an
expression, the name being the symbol name to be found at link
time. For the first version of the code generator, all variables
and function parameters will be assumed to be double-precision
floats.