Add basic command line flags to main
The --help/-h --version and --compile/-S command line flags now
work.

The help text is currently just the command line flags--the flags
themselves have descriptions but the basic text to describe what
Funk is and how to use it is just a placeholder.

Specifying -S currently just causes the LLVM IR to be
printed to standard output, since the --output option isn't hooked
up yet.

Also, this code could use some cleanup. It's not terrible but it's
not beautiful either.