(Refactor) Remove unused type vars in Funk.Names
Funk.Names exported 3 types: "UnresolvedName a" "ResolvedName a"
and "RawName", where RawName was simply an alias for
"UnresolvedName ()". The "a" type parameter is only ever specified
as () everywhere any of these classes are used and I don't remember
what it was for (And didn't write any enlightening comments
either). Possibly I got confused with the type parameter that the
AST types take.

Anyway, this commit removes the type parameter from UnresolvedName
and ResolvedName. This makes RawName simply a synonym for
UnresolvedName so I removed RawName as well.

This is a pure refactor which shouldn't cause any change in
functionality.